Add CHIP-8 Screen
Implement the CHIP-8 screen, also implement the functions required in the main function to allow rendering of the screen. Memory was modifed to allow the loading of sprites from memory. Issue with the memory was also resolved where only 4 bytes of a block would be loaded. Tested rendering the default character set directly from memory.