Partial Implementation Of Opcodes
Opcodes implemented upto the 0x8000 opcodes. There is an issue with
loading to and from V registers that are > 9 e.g. when the hex value
is A, B, C, D, F. This need to be invenstigated and resolved.

This seems to be happening most often in the 0x8xy0 opcode. This may
me unique to this opcode it has not been seen in others.