Complete Extended Opcodes for 0x8000
The last of the opcodes for 0x8000 opcodes were completed and tested
manually, using very masic observations. Full testing of the opcodes
will be needed once all opcodes are implemented.

Chip8_Stack debug messages where updated to display the index values
are pushed/ popped to as an int not a char, this way they show the
shack index correctly.