Add UnitTestChip8Emulator To The Solution
The Unit Test has been added to the solution for a few reasons. As,
I have been keen to learn about unit tests for some time but it has
also been added to help identify the issues raised in Iss#2 as at
the current time it is very hard to track down which instruction /
component has the issue.

At this time only one example test has been written for the
Chip8_Stack, this was just a proof of concept. An Class has been
created for every main component in the emulator the aim is to wirte
a test for every function and opcode to ensure they operate as
expected.